firebase-cloud-function/functions/requirements.txt
firebase-functions
firebase-admin
requests
google-generativeai
textblob
python-dotenv
yfinance